improve stripe connect setup ui and add 3 different states
